adding event listener after posting result, deleting useless files, adding betOdds for bet model

+++ src/main/java/com/dominik/typer/service/userpersistence/UserPersistence.java
@@ -2,7 +2,6 @@ package com.dominik.typer.service.userpersistence;
 
 import com.dominik.typer.model.User;
 
-import java.math.BigDecimal;
 import java.util.List;
 import java.util.Optional;
 
@@ -15,5 +14,5 @@ public interface UserPersistence {
     void deleteUserById(Integer id);
     void updateUserById(Integer id, User user);
     Optional<User> getUserByUsername(String username);
-    void updateBalance(Integer id, BigDecimal balance);
+    void updateBalance(Integer id, Double balance);
 }
